# gpt-actions.openapi.yaml
openapi: 3.1.0
info:
  title: Windows LocalControl GPT Actions
  description: Private Custom GPT action schema for a single-user Windows LocalControl
    bridge. Full-control execution mode is enabled; use /execution/logs to retrieve
    command and terminal output.
  version: 0.1.0
paths:
...
  /terminal/sessions:
    post:
      tags:
      - terminal
      summary: Terminal Create Session
      operationId: terminal_create_session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalSessionCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSessionInfo'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /terminal/sessions/list:
    post:
      tags:
      - terminal
      summary: Terminal List Sessions
      operationId: terminal_list_sessions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalSessionListRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSessionListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /terminal/sessions/{session_id}:
    get:
      tags:
      - terminal
      summary: Terminal Get Session
      operationId: terminal_get_session
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSessionInfo'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /terminal/sessions/{session_id}/exec:
    post:
      tags:
      - terminal
      summary: Terminal Exec
      operationId: terminal_exec
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalExecRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalExecResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /terminal/sessions/{session_id}/stdin:
    post:
      tags:
      - terminal
      summary: Terminal Stdin
      operationId: terminal_stdin
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalStdinRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalStdinResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /terminal/sessions/{session_id}/events:
    post:
      tags:
      - terminal
      summary: Terminal Events
      operationId: terminal_events
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalEventsRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalEventsResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /terminal/sessions/{session_id}/terminate:
    post:
      tags:
      - terminal
      summary: Terminal Terminate
      operationId: terminal_terminate
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalTerminateResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: false
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
      additionalProperties: false
      type: object
      required:
      - code
      - message
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TerminalEvent:
      properties:
        event_id:
          type: integer
          title: Event Id
        timestamp:
          type: string
          title: Timestamp
        stream:
          type: string
          enum:
          - command
          - stdin
          - stdout
          - stderr
          - system
          title: Stream
        text:
          type: string
          title: Text
      additionalProperties: false
      type: object
      required:
      - event_id
      - timestamp
      - stream
      - text
      title: TerminalEvent
    TerminalEventsRequest:
      properties:
        after_event_id:
          type: integer
          minimum: 0.0
          title: After Event Id
          default: 0
        max_events:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Max Events
          default: 100
      additionalProperties: false
      type: object
      title: TerminalEventsRequest
    TerminalEventsResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        status:
          type: string
          title: Status
        events:
          items:
            $ref: '#/components/schemas/TerminalEvent'
          type: array
          title: Events
        next_event_id:
          type: integer
          title: Next Event Id
      additionalProperties: false
      type: object
      required:
      - session_id
      - status
      - events
      - next_event_id
      title: TerminalEventsResponse
    TerminalExecRequest:
      properties:
        command:
          type: string
          minLength: 1
          title: Command
        include_secrets:
          type: boolean
          title: Include Secrets
          default: false
      additionalProperties: false
      type: object
      required:
      - command
      title: TerminalExecRequest
    TerminalExecResponse:
      properties:
        session:
          $ref: '#/components/schemas/TerminalSessionInfo'
        command_id:
          type: string
          title: Command Id
      additionalProperties: false
      type: object
      required:
      - session
      - command_id
      title: TerminalExecResponse
    TerminalSessionCreateRequest:
      properties:
        shell:
          type: string
          enum:
          - powershell
          - cmd
          title: Shell
          default: powershell
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
        cwd:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwd
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        env:
          additionalProperties:
            type: string
          type: object
          title: Env
      additionalProperties: false
      type: object
      title: TerminalSessionCreateRequest
    TerminalSessionInfo:
      properties:
        session_id:
          type: string
          title: Session Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        shell:
          type: string
          title: Shell
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
        cwd:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwd
        status:
          type: string
          title: Status
        created_at:
          type: string
          title: Created At
        last_active_at:
          type: string
          title: Last Active At
        process_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Process Id
        command_count:
          type: integer
          title: Command Count
      additionalProperties: false
      type: object
      required:
      - session_id
      - shell
      - cwd
      - status
      - created_at
      - last_active_at
      - command_count
      title: TerminalSessionInfo
    TerminalSessionListRequest:
      properties:
        include_exited:
          type: boolean
          title: Include Exited
          default: true
      additionalProperties: false
      type: object
      title: TerminalSessionListRequest
    TerminalSessionListResponse:
      properties:
        sessions:
          items:
            $ref: '#/components/schemas/TerminalSessionInfo'
          type: array
          title: Sessions
      additionalProperties: false
      type: object
      required:
      - sessions
      title: TerminalSessionListResponse
    TerminalStdinRequest:
      properties:
        input:
          type: string
          title: Input
      additionalProperties: false
      type: object
      required:
      - input
      title: TerminalStdinRequest
    TerminalStdinResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        bytes_written:
          type: integer
          title: Bytes Written
      additionalProperties: false
      type: object
      required:
      - session_id
      - bytes_written
      title: TerminalStdinResponse
    TerminalTerminateResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        status:
          type: string
          title: Status
      additionalProperties: false
      type: object
      required:
      - session_id
      - status
      title: TerminalTerminateResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://YOUR-RESERVED-NGROK-DOMAIN.ngrok-free.app